bug fixed

+++ gnn/neighbor_sampler.py
@@ -17,4 +17,4 @@ def sample_neighs(G, nodes, sample_num=None, self_loop=False, shuffle=True):
             samp_neighs = [list(np.random.permutation(x)) for x in samp_neighs]
     else:
         samp_neighs = neighs
-    return np.asarray(samp_neighs), np.asarray(list(map(len, samp_neighs)))
+    return np.asarray(samp_neighs, dtype=np.int64), np.asarray(list(map(len, samp_neighs)), dtype=np.int64)
